Se agrega capa de Servicio, Controlador y repositorio para User

+++ src/main/resources/db/migration/V1__crearTablaUsers.sql
@@ -1,6 +1,6 @@
 
 CREATE TABLE user(
-    id int not null auto_increment,
+    id bigint not null auto_increment,
     nombre varchar(60) not null,
     email varchar(30) not null,
     password varchar(50) not null,
